+++ plone/recipe/codeanalysis/i18ndude.py
@@ -16,9 +16,18 @@ class I18NDude(Analyser):
         if files:
             cmd.append(self.options.get('i18ndude-bin') or '')
             cmd.append('find-untranslated')
+            if self.nosummary:
+                cmd.append('--nosummary')
             cmd.extend(files)
 
         return cmd
+
+    @property
+    def nosummary(self):
+        """The report will contain only the errors for each file."""
+        return I18NDude.normalize_boolean(
+            self.get_prefixed_option('no-summary')
+        )
 
 
 def console_script(options):